escalation/requirements-app.txt
click
colorama
cython
flask
flask_sqlalchemy
Flask-Migrate
gunicorn
jinja2
jsonschema
markupsafe
numpy
pandas
pathvalidate
pg8000
psycopg2-binary
plotly
pyyaml
requests
setuptools
six
sqlacodegen
sqlalchemy
tableschema
werkzeug
wheel